pyproject.toml
[tool.poetry]
name = "usermanagementservice"
version = "0.1.0"
description = "User management service"
authors = ["Alexander Rusakevich <mr.alexander.rusakevich@gmail.com>"]
license = "GPLv3"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.12"
alembic = "^1.16.1"
uvicorn = "^0.34.2"
redis = "^6.1.0"
pydantic-settings = "^2.9.1"
asyncpg = "^0.30.0"
psycopg2-binary = "^2.9.10"
aioboto3 = "^14.3.0"
types-aioboto3 = { extras = ["full"], version = "^14.3.0" }
python-jose = "^3.4.0"


[tool.poetry.group.dev.dependencies]
poethepoet = "^0.34.0"
types-aioboto3 = { extras = ["full"], version = "^14.3.0" }

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poe.tasks]
serve.script = "app.main:main"
dev.script = "app.main:dev"
docker-up = "sudo docker-compose up --build app"
docker-down = "sudo docker-compose down"
